4.12

+++ RPG_Frame/include/Math2.h
@@ -4,11 +4,11 @@
 namespace GameFrame {
 	class Vector2 {
 	public:
-		float x;
-		float y;
+		int x;
+		int y;
 
 		Vector2();
-		Vector2(float X, float Y);
+		Vector2(int X, int Y);
 
 		static const Vector2 up;
 		static const Vector2 down;
@@ -19,7 +19,7 @@ namespace GameFrame {
 		friend Vector2 operator-(const Vector2& a, const Vector2& b);
 		friend Vector2 operator*(const Vector2& a, const Vector2& b);
 		friend Vector2 operator*(const Vector2& a, float b);
-		friend Vector2 operator*(float a, const Vector2& b);
+		friend Vector2 operator*(int a, const Vector2& b);
 	};
 }
 #endif //
